feat(passport): Revocation of smart contract wallet scope
