+++ platform/address/package.json
@@ -47,7 +47,7 @@
     "@proofzero/utils": "workspace:*",
     "@trpc/server": "10.8.1",
     "@types/node": "18.15.3",
-    "@zerodevapp/sdk": "3.0.15",
+    "@zerodevapp/sdk": "3.1.33",
     "do-proxy": "1.3.3",
     "jose": "4.11.0",
     "remix-auth-google": "1.2.0",
