+++ platform/address/src/context.ts
@@ -43,6 +43,7 @@ interface CreateInnerContextOptions
   DELAY_BETWEEN_REGENERATION_ATTEMPTS_IN_MS: number
   REGENERATION_COOLDOWN_PERIOD_IN_MS: number
   MAX_ATTEMPTS_TIME_PERIOD_IN_MS: number
+  MUMBAI_PROVIDER_URL: string
 
   SECRET_ZERODEV_PROJECTID: string
