+++ platform/address/src/types.ts
@@ -25,6 +25,7 @@ export interface Environment {
   DELAY_BETWEEN_REGENERATION_ATTEMPTS_IN_MS: number
   REGENERATION_COOLDOWN_PERIOD_IN_MS: number
   MAX_ATTEMPTS_TIME_PERIOD_IN_MS: number
+  MUMBAI_PROVIDER_URL: string
 
   SECRET_ZERODEV_PROJECTID: string
